backend/analytics-service/requirements.txt
Flask==3.0.0
redis==5.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
PyJWT==2.8.0